+++ src/test/java/org/dvsa/testing/framework/Journeys/licence/GovSignInJourney.java
@@ -53,16 +53,29 @@ public class GovSignInJourney extends BasePage {
             waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
         }
 
-        if (Objects.requireNonNull(navigate().getCurrentUrl()).contains("enter-email")) {
-            waitAndEnterText("email", SelectorType.ID, signInUsername);
-            waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
-            waitAndEnterText("password", SelectorType.ID, signInPassword);
-            waitAndClick("//button[@type='Submit']", SelectorType.XPATH);
-            waitAndEnterText("code", SelectorType.ID, authCode);
-            waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
+        if (Objects.requireNonNull(navigate().getCurrentUrl()).contains("https://identity.integration.account.gov.uk/ipv/page/page-ipv-reuse")) {
+            waitAndClick("//*[@id='submitButton']", SelectorType.XPATH);
+        } else if (Objects.requireNonNull(navigate().getCurrentUrl()).contains("enter-email")) {
+            if (isTextPresent("You have already proved your identity")) {
+                waitAndClick("//*[@id='submitButton']", SelectorType.XPATH);
+            } else {
+                waitAndEnterText("email", SelectorType.ID, signInUsername);
+                waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
+                waitAndEnterText("password", SelectorType.ID, signInPassword);
+                waitAndClick("//button[@type='Submit']", SelectorType.XPATH);
+                waitAndEnterText("code", SelectorType.ID, authCode);
+                waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
+                if (isTextPresent("You have already proved your identity")) {
+                    waitAndClick("//*[@id='submitButton']", SelectorType.XPATH);
+                } else {
+                    waitAndEnterText("code", SelectorType.ID, authCode);
+                    waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
+                }
+            }
+        }
 
-            waitAndEnterText("code", SelectorType.ID, authCode);
-            waitAndClick("//*[contains(text(),'Continue')]", SelectorType.XPATH);
+        if (Objects.requireNonNull(navigate().getCurrentUrl()).contains("https://identity.integration.account.gov.uk/ipv/page/page-ipv-reuse")) {
+            waitAndClick("//*[@id='submitButton']", SelectorType.XPATH);
         }
 
         if (Objects.requireNonNull(navigate().getCurrentUrl()).contains("sign-in-or-create")) {